Move some utils code around so that graphics do not need to depend on painters.
git-svn-id: https://svn.java.net/svn/swingx~svn/trunk@4076 1312f61e-266d-0410-97fa-c3b71232c9ac

+++ swingx-common/src/main/java/org/jdesktop/swingx/graphics/ShapeUtils.java
@@ -34,58 +34,57 @@ import java.awt.geom.Point2D;
 import java.awt.geom.Rectangle2D;
 import java.awt.image.BufferedImage;
 
-
 /**
- *
+ * 
  * @author joshy
  */
 public final class ShapeUtils {
-    
+
     /** Creates a new instance of ShapeUtils */
-    private ShapeUtils() { }
-    
+    private ShapeUtils() {
+    }
+
     public static Shape generatePolygon(int sides, int outsideRadius, boolean normalize) {
         return generatePolygon(sides, outsideRadius, 0, normalize);
     }
-    
-    public static Shape generatePolygon(int sides, int outsideRadius, int insideRadius, boolean normalize) {
-        Shape shape = generatePolygon(sides,outsideRadius,insideRadius);
-        if(normalize) {
+
+    public static Shape generatePolygon(int sides, int outsideRadius, int insideRadius,
+            boolean normalize) {
+        Shape shape = generatePolygon(sides, outsideRadius, insideRadius);
+        if (normalize) {
             Rectangle2D bounds = shape.getBounds2D();
             GeneralPath path = new GeneralPath(shape);
-            shape = path.createTransformedShape(
-                    AffineTransform.getTranslateInstance(-bounds.getX(), -bounds.getY()));
+            shape = path.createTransformedShape(AffineTransform.getTranslateInstance(
+                    -bounds.getX(), -bounds.getY()));
         }
         return shape;
     }
-    
-    
+
     public static Shape generatePolygon(int sides, int outsideRadius, int insideRadius) {
-        
-        if(sides < 3) {
-            return new Ellipse2D.Float(0,0,10,10);
+        if (sides < 3) {
+            return new Ellipse2D.Float(0, 0, 10, 10);
         }
-        
+
         AffineTransform trans = new AffineTransform();
         Polygon poly = new Polygon();
-        for(int i=0; i<sides; i++) {
-            trans.rotate(Math.PI*2/(float)sides/2);
-            Point2D out = trans.transform(new Point2D.Float(0,outsideRadius),null);
-            poly.addPoint((int)out.getX(), (int)out.getY());
-            trans.rotate(Math.PI*2/(float)sides/2);
-            if(insideRadius > 0) {
-                Point2D in = trans.transform(new Point2D.Float(0,insideRadius),null);
-                poly.addPoint((int)in.getX(), (int)in.getY());
+        for (int i = 0; i < sides; i++) {
+            trans.rotate(Math.PI * 2 / sides / 2);
+            Point2D out = trans.transform(new Point2D.Float(0, outsideRadius), null);
+            poly.addPoint((int) out.getX(), (int) out.getY());
+            trans.rotate(Math.PI * 2 / sides / 2);
+            if (insideRadius > 0) {
+                Point2D in = trans.transform(new Point2D.Float(0, insideRadius), null);
+                poly.addPoint((int) in.getX(), (int) in.getY());
             }
         }
-        
+
         return poly;
     }
 
     public static Shape generateShapeFromText(Font font, char ch) {
         return generateShapeFromText(font, String.valueOf(ch));
     }
-    
+
     public static Shape generateShapeFromText(Font font, String string) {
         BufferedImage img = GraphicsUtilities.createCompatibleTranslucentImage(1, 1);
         Graphics2D g2 = img.createGraphics();
@@ -97,7 +96,7 @@ public final class ShapeUtils {
             return shape;
         } finally {
             g2.dispose();
-       }
+        }
     }
 
     /**
@@ -109,19 +108,19 @@ public final class ShapeUtils {
      *            the graphics object to update
      * @param clip
      *            a new clipping region to add to the graphics clip.
-     * @return the current clipping region of the supplied graphics object. This may return {@code
-     *         null} if the current clip is {@code null}.
+     * @return the current clipping region of the supplied graphics object. This may return
+     *         {@code null} if the current clip is {@code null}.
      * @throws NullPointerException
      *             if any parameter is {@code null}
      */
     public static Shape mergeClip(Graphics g, Shape clip) {
         Shape oldClip = g.getClip();
-        if(oldClip == null) {
+        if (oldClip == null) {
             g.setClip(clip);
             return null;
         }
         Area area = new Area(oldClip);
-        area.intersect(new Area(clip));//new Rectangle(0,0,width,height)));
+        area.intersect(new Area(clip));// new Rectangle(0,0,width,height)));
         g.setClip(area);
         return oldClip;
     }
